% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Warszawa.R
\docType{data}
\encoding{UTF-8}
\name{Warszawa}
\alias{Warszawa}
\title{Polygons for 18 districts of Warsaw}
\format{Object of class "spatialPolygonsDataFrame" with some variables to be
described later.}
\source{
Centralny Ośrodek Dokumentacji Geodezyjnej i Kartograficznej (CODGiK)
\url{codgik.gov.pl}
}
\description{
Administrative map of Warsaw
}
\keyword{datasets}

