% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getShape.R
\name{getShape}
\alias{getShape}
\title{Wrapper for reading shape files}
\usage{
getShape(nam = c("voiv", "polska"), ...)
}
\arguments{
\item{nam}{character, base name of the shape file to be read}

\item{...}{other arguments passed to \code{\link{readShapePoly}}}
}
\value{
Object of class "\linkS4class{SpatialPolygonsDataFrame}".
}
\description{
This is a simple wrapper around \code{\link{readShapePoly}} for importing
shape files contained within package \pkg{mapoland}.
}
\examples{
s <- getShape("voiv")
# plot(s)
}
\seealso{
\code{\link{readShapePoly}}, \link{mapoland}
}

