\name{mapoland-package}
\alias{mapoland-package}
\alias{mapoland}
\docType{package}
\title{Maps of Poland}
\description{
This package contains data and tools for drawing an maps of
Poland.
}
\details{
This package enables drawing data on administrative maps of Poland. It includes
data on the voivodship. Future releases will include county and municipality
levels.

See \code{help(package="mapoland")} for the complete list of functions.

We have invested some time and effort in creating this package. We hope that it
will be usefull to you. Please cite it whenever you use it for data analysis or
other tasks. See \code{citation(package="mapoland")} for the proper information
and format.
}
\author{
Michal Bojanowski and Tomasz W. Jerzynski

Maintainer: Michal Bojanowski \email{michal2992@gmail.com}
}
\seealso{
On CRAN you can also find other packages for drawing maps, e.g. \pkg{maps}.
}
\examples{
# TODO
}
\keyword{package}
