% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/voivData.R
\docType{data}
\name{voivData}
\alias{voivData}
\title{Data on Polish voivodships}
\format{A data frame with 16 observations on the following 4 variables.
\describe{
\item{teryt}{numeric, area code from "TERYT": official Polish area
classification}
\item{area}{numeric, area in square kilometers}
\item{perimeter}{numeric, voivodship perimeter in #' kilometers}
\item{pop2001}{numeric, voivodship population in 2001}
}}
\source{
Official statistics: \url{http://www.stat.gov.pl}
TERYT database: \url{http://www.stat.gov.pl/broker/access/index.jspa}
}
\description{
Some extra data on Polish voivodships.
}
\examples{
if( require(sp) ) {
# areas of voivodships
a <- structure( voivData$area, names=rownames(voivData))
s <- getShape("voiv")
k <- coordinates(s)
# colors
x <- a - min(a)
kol <- x / max(x)
plot(s, col=gray(kol))
text( coordinates(s), labels=voivNames(), col=gray((1-kol)>.5),
    cex=.8)
}
}
\keyword{datasets}

